Use java-test-fixtures plugin to optimize build
The junit-jupiter-api subproject now uses the new java-test-fixtures
Gradle plugin to expose a testFixtures jar. The `TrackLogRecords`
annotation and its accompanying extension are moved from
junit-jupiter-engine. Prior to this commit, any change to tests in
junit-jupiter-engine caused tests in junit-vintage-engine and
platform-tests to be rerun. Now, they are only rerun if
`TrackLogRecords` changes.